backend/app/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
pydantic-settings
passlib[bcrypt]
python-jose
alembic
shapely
geojson
requests
redis
